refactor: move cli doctor modules into cli/doctor/ subpackage
Convert _doctor.py facade to doctor/__init__.py and move all 9
_doctor_*.py sub-modules alongside it.
