+++ src/autoskillit/cli/doctor/_doctor_fleet.py
