+++ src/autoskillit/cli/doctor/_doctor_install.py
