+++ src/autoskillit/cli/doctor/_doctor_mcp.py
